BRS-1139: moving writeReservation utility to common area.
